AIM-182 document text extract function refactoring.
